refactored codec and patch initialisation, implemented variable block size

+++ Source/ApplicationSettings.h
@@ -31,11 +31,10 @@ public:
   I2SProtocol audio_codec_protocol;
   uint32_t audio_samplingrate;
   uint8_t audio_bitdepth;
-  uint32_t audio_blocksize;
+  uint16_t audio_blocksize;
   uint8_t patch_green;
   uint8_t patch_red;
   uint8_t patch_mode;
-  /* bool patch_midi_mode; */
   uint8_t midi_channel;
   uint8_t inputGainLeft;
   uint8_t inputGainRight;
